feat: Fix TypeScript type errors from yarn typecheck

Plan steps:
1. [modify] Remove redundant instanceof String checks from leaguecard.tsx - fix lines 26 and 31 by replacing 'typeof x === 'string' || x instanceof String' with just 'typeof x === 'string'' (files: components/leaguecard.tsx)
2. [modify] Remove redundant instanceof String checks from statcard.tsx - fix lines 33 and 38 by replacing 'typeof x === 'string' || x instanceof String' with just 'typeof x === 'string'' (files: components/statcard.tsx)
3. [modify] Remove invalid 'id' property from Chart.js dataset in statcard.tsx - delete 'id: 1' from the dataset object at line 78 (files: components/statcard.tsx)
4. [modify] Fix SessionProvider JSX type compatibility in pages/_app.tsx by adding proper type annotation and ensuring next-auth/react is correctly imported (files: pages/_app.tsx)
5. [modify] Fix next-auth import path in pages/api/auth/[...nextauth].ts - change import from 'next-auth/providers' to use correct next-auth v4 types (files: pages/api/auth/[...nextauth].ts)
6. [modify] Fix Buffer to Uint8Array type mismatch in utils/yahooData.ts - add proper type casting for all zlib.gunzip calls (lines 41-42, 92-93, 143-144, 206-207) to convert Buffer to Uint8Array (files: utils/yahooData.ts)
7. [modify] Fix unknown type in pages/api/teams.ts - properly type the return value from getTeams() to match ResponseData or add type guard to ensure compatibility (files: pages/api/teams.ts)

+++ utils/yahooData.ts
@@ -40,7 +40,7 @@ export const getTeams = async (req: NextApiRequest): Promise<unknown> => {
         response.on('end', function () {
           const buffer = Buffer.concat(chunks);
           zlib.gunzip(buffer, (_err, dezipped) => {
-            parser.parseString(dezipped.toString(), function (_err, result) {
+            parser.parseString((dezipped as Uint8Array).toString(), function (_err, result) {
               games = (result as Record<string, unknown>).fantasy_content;
               resolve(games);
             });
@@ -91,7 +91,7 @@ export const getLeagueTeams = async (
         response.on('end', function () {
           const buffer = Buffer.concat(chunks);
           zlib.gunzip(buffer, (_err, dezipped) => {
-            parser.parseString(dezipped.toString(), function (_err, result) {
+            parser.parseString((dezipped as Uint8Array).toString(), function (_err, result) {
               league = (result as Record<string, unknown>).fantasy_content;
               resolve(league);
             });
@@ -142,7 +142,7 @@ export const getLeagueSettings = async (
         response.on('end', function () {
           const buffer = Buffer.concat(chunks);
           zlib.gunzip(buffer, (_err, dezipped) => {
-            parser.parseString(dezipped.toString(), function (_err, result) {
+            parser.parseString((dezipped as Uint8Array).toString(), function (_err, result) {
               league = (result as Record<string, unknown>).fantasy_content;
               resolve(league);
             });
@@ -205,7 +205,7 @@ export const getWeekStats = async (
         response.on('end', function () {
           const buffer = Buffer.concat(chunks);
           zlib.gunzip(buffer, (_err, dezipped) => {
-            parser.parseString(dezipped.toString(), function (_err, result) {
+            parser.parseString((dezipped as Uint8Array).toString(), function (_err, result) {
               stats = (result as Record<string, unknown>).fantasy_content;
               resolve(stats);
             });
